Move markdown parsing functions into own exported file.
