+++ _data/theme_hash.yml
@@ -1,1 +1,1 @@
-2bf85ef194749423bbcd
+5418c2bebacc724f777b
