+++ assets/js/theme/emergency/createAlertsHtml.js
@@ -2,6 +2,7 @@
 // ==================================================================================
 // This exported module requires you pass the function the `response` object from the API call,
 // and `resolve()` from the Promise that the function call is wrapped in (see './campusAlertsSheetsAPI.js')
+import parseMarkdownToHTML from './parseMarkdownToHTML.js';
 
 const CAMPUS_ALERTS_DIV_ID_STRING = 'emergencyAlerts';  // ID of the div to house campus alerts - already built into the page.
 const BOOTSTRAP_CONTAINER_CLASS = 'container';  // Class used in Bootstrap 4
@@ -26,74 +27,6 @@ function appendElementWithNewNode(el, nodeType, classArr) {
   classArrIsNotIterable ? newNode.classList.add(classArr) : addClassesToEl(newNode, classArr);
   el.appendChild(newNode);
   return newNode;
-}
-
-// =========================================================================== //
-//                                                                             //
-//                                                                             //
-//                                                                             //
-//     TODO:   Breakout markdown parsing functions into its own export         //
-//             Then, import into this file.                                    //
-//                                                                             //
-//                                                                             //
-//                                                                             //
-// =========================================================================== //
-
-
-function replaceRegex(string, regex, replacement) {
-  const newString = string.replace(regex, replacement);
-  return newString;
-}
-
-function escapeCharacters(string, escapeOption) {
-  const escapedCharactersObject = {
-    '\\*': '__asterisk__',
-    '\\_': '__underscore__',
-    '\\[': '__openBracket__',
-    '\\]': '__closeBracket__',
-    '\\(': '__openParenthesis__',
-    '\\)': '__closeParenthesis__'
-  }
-
-  for (let char in escapedCharactersObject) {
-    if (escapedCharactersObject.hasOwnProperty(char)) {
-      escapeOption === true ? string = replaceRegex(string, char, escapedCharactersObject[char])
-      : escapeOption === false ? string = replaceRegex(string, escapedCharactersObject[char], char.replace(/^\\/g, ''))
-      : null;
-    }
-  }
-  return string;
-}
-
-function createAnchorElements(string) {
-  return string = string.replace(/\[(?<linkText>[^\]]*)\]\((?<linkHref>[^\)]*)\)/g, '<a href="$<linkHref>" target="_blank" rel="noopener noreferrer">$<linkText></a>');
-}
-
-function createInlineElements(string) {
-  const inlineElementObject = {
-    'strong': /\*\*([^\*]*)\*\*/g,
-    'em': /(?<!_)_([^_]*)(?<!_)_/g // That's some pretty intense RegEx right there: "Negative lookbehind" to omit the escaped characters
-  }
-  for (var el in inlineElementObject) {
-    if (inlineElementObject.hasOwnProperty(el)) {
-      string = string.replace(inlineElementObject[el], '<' + el + '>$1</' + el + '>' );
-    }
-  }
-  return string;
-}
-
-function createParagraphElements(string) {
-   return string = string.replace(/(?<!$)^(.*)(?<!^)$/gm, '<p class="typography__alert">$1</p>');
-}
-
-function parseMarkdownToHTML(string) {
-  const escapedString = escapeCharacters(string, true);
-  const stringWithInlineElements = createInlineElements(escapedString);
-  const stringWithAnchorElements = createAnchorElements(stringWithInlineElements);
-  const stringWithParagraphElements = createParagraphElements(stringWithAnchorElements);
-  const unescapedString = escapeCharacters(stringWithParagraphElements, false);
-  //console.log(unescapedString);
-  return unescapedString;
 }
 
 function createHTML(SHEET_DATA) { // Finally....after all those checks....we get to do something!
